Rename `ui-constants.ts` to `ui.ts`
