+++ src/loading-scene.ts
@@ -22,7 +22,7 @@ import { initVouchers } from "#app/system/init-vouchers";
 import { DEFAULT_LANGUAGE_KEY } from "#app/system/settings/supported-languages";
 import { timedEventManager } from "#app/timed-event-manager";
 import { isMobile } from "#app/touch-controls";
-import { CANVAS_SCALE, GAME_HEIGHT, GAME_WIDTH, TEMP_SCALE_ADJUSTMENT } from "#app/constants/ui-constants";
+import { CANVAS_SCALE, GAME_HEIGHT, GAME_WIDTH, TEMP_SCALE_ADJUSTMENT } from "#app/constants/ui";
 import { getWindowVariantSuffix } from "#app/ui/ui-theme";
 import { getEnumKeys, getEnumValues } from "#app/utils";
 import { BiomeId } from "#enums/biome-id";
