+++ src/phases/end-card-phase.ts
@@ -5,7 +5,7 @@ import { TextStyle } from "#enums/text-style";
 import { PlayerGender } from "#enums/player-gender";
 import i18next from "i18next";
 import { settings } from "#app/system/settings/settings-manager";
-import { GAME_WIDTH, GAME_HEIGHT } from "#app/constants/ui-constants";
+import { GAME_WIDTH, GAME_HEIGHT } from "#app/constants/ui";
 import { PhaseId } from "#enums/phase-id";
 
 /**
