+++ src/pipelines/sprite.ts
@@ -1,5 +1,5 @@
 import FieldSpritePipeline from "#app/pipelines/field-sprite";
-import { CANVAS_SCALE } from "#app/constants/ui-constants";
+import { CANVAS_SCALE } from "#app/constants/ui";
 
 const spriteFragShader = `
 #ifdef GL_FRAGMENT_PRECISION_HIGH
