+++ src/ui/components/achv-bar.ts
@@ -3,7 +3,7 @@ import { Achievement } from "#app/system/achievements";
 import type { Voucher } from "#app/system/voucher";
 import { addTextObject } from "#app/ui/text/text-utils";
 import { TextStyle } from "#enums/text-style";
-import { GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui-constants";
+import { GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui";
 
 export class AchvBar extends Phaser.GameObjects.Container {
   private defaultWidth: number;
