+++ src/ui/components/bgm-bar.ts
@@ -4,7 +4,7 @@ import i18next from "i18next";
 import { formatText } from "#app/utils";
 import { globalScene } from "#app/global-scene";
 import { settings } from "#app/system/settings/settings-manager";
-import { TEXT_SCALE } from "#app/constants/ui-constants";
+import { TEXT_SCALE } from "#app/constants/ui";
 
 const hiddenX = -150;
 const shownX = 0;
