+++ src/ui/components/saving-icon.ts
@@ -1,6 +1,6 @@
 import { globalScene } from "#app/global-scene";
 import { fixedNumber } from "#app/utils";
-import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui-constants";
+import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui";
 
 export class SavingIcon extends Phaser.GameObjects.Container {
   private icon: Phaser.GameObjects.Sprite;
