+++ src/ui/components/timed-event-display.ts
@@ -1,5 +1,5 @@
 import { globalScene } from "#app/global-scene";
-import { GAME_HEIGHT } from "#app/constants/ui-constants";
+import { GAME_HEIGHT } from "#app/constants/ui";
 import { TextStyle } from "#enums/text-style";
 import type { TimedEvent } from "#app/@types/TimedEvent";
 import i18next from "i18next";
