+++ src/ui/handlers/challenges-select-ui-handler.ts
@@ -1,7 +1,7 @@
 import type { Challenge } from "#app/data/challenge";
 import { globalScene } from "#app/global-scene";
 import { SelectStarterPhase } from "#app/phases/select-starter-phase";
-import { GAME_HEIGHT, GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui-constants";
+import { GAME_HEIGHT, GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui";
 import { addBBCodeTextObject, addTextObject } from "#app/ui/text/text-utils";
 import { addWindow } from "#app/ui/ui-theme";
 import { Button } from "#enums/buttons";
