+++ src/ui/handlers/fight-ui-handler.ts
@@ -5,7 +5,7 @@ import type { PokemonMove } from "#app/field/pokemon-move";
 import { globalScene } from "#app/global-scene";
 import type { CommandPhase } from "#app/phases/command-phase";
 import { settings } from "#app/system/settings/settings-manager";
-import { GAME_WIDTH } from "#app/constants/ui-constants";
+import { GAME_WIDTH } from "#app/constants/ui";
 import { MoveInfoOverlay } from "#app/ui/components/move-info-overlay";
 import { addTextObject, setTextColor } from "#app/ui/text/text-utils";
 import { fixedNumber, leftPad } from "#app/utils";
