+++ src/ui/handlers/form-change-scene-ui-handler.ts
@@ -1,7 +1,7 @@
 import { globalScene } from "#app/global-scene";
 import type { EvolutionPhase } from "#app/phases/evolution-phase";
 import { settings } from "#app/system/settings/settings-manager";
-import { GAME_HEIGHT, GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui-constants";
+import { GAME_HEIGHT, GAME_WIDTH, TEXT_SCALE } from "#app/constants/ui";
 import { addTextObject } from "#app/ui/text/text-utils";
 import { Button } from "#enums/buttons";
 import { PhaseId } from "#enums/phase-id";
