+++ src/ui/handlers/modifier-select-ui-handler.ts
@@ -7,7 +7,7 @@ import { getPlayerShopModifierTypeOptionsForWave, TmModifierType } from "#app/mo
 import Overrides from "#app/overrides";
 import { settings } from "#app/system/settings/settings-manager";
 import { handleTutorial } from "#app/tutorial";
-import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui-constants";
+import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui";
 import { MoveInfoOverlay } from "#app/ui/components/move-info-overlay";
 import { addTextObject, getModifierTierTextTint, setTextColor } from "#app/ui/text/text-utils";
 import { formatMoney, isNullOrUndefined, NumberHolder } from "#app/utils";
