+++ src/ui/handlers/title-ui-handler.ts
@@ -2,7 +2,7 @@ import { getSplashMessages } from "#app/data/splash-messages";
 import { globalScene } from "#app/global-scene";
 import { api } from "#app/plugins/api/api";
 import { timedEventManager } from "#app/timed-event-manager";
-import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui-constants";
+import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui";
 import { TimedEventDisplay } from "#app/ui/components/timed-event-display";
 import type { OptionSelectModeConfig } from "#app/ui/interfaces/option-select-config";
 import { addTextObject } from "#app/ui/text/text-utils";
